gooseberry/tests/test_db_access.py
# This is a simplified example. Adjust based on your testing strategy.

import pytest
from gooseberry.db.db_access import DBAccess

@pytest.fixture(scope="module")
def db_access():
    # Setup code to initialize DBAccess with test database credentials
    db = DBAccess(uri="bolt://localhost:7687", user="neo4j", password="test")
    yield db
    # Teardown code to close connection, clean up test data, etc.
    db.close()

def test_create_node(db_access):
    properties = {"name": "TestNode"}
    node = db_access.create_node("TestLabel", properties)
    assert node is not None
    # Additional assertions to verify the node's properties, label, etc.
